add tests for deprecated alias
